dtype refactor

+++ op_tests/triton_tests/attention/test_sage_attn.py
@@ -202,7 +202,7 @@ def test_sage(
     NUM_K_HEADS: int,
     HEAD_SZ: int,
     layout: str,
-    dtype=torch.float16,
+    dtype=torch.bfloat16,
 ):
     torch.cuda.empty_cache()
